requirements.txt
fastap
sqlalchemy
pydantic
pymysql
passlib
python-jose
uvicorn
bcrypt